+++ index.html
@@ -13,7 +13,7 @@
         <title>Vite + React</title>
     </head>
     <body>
-        <div id="overlays"></div>
+        <div id="overlays" class="flex"></div>
         <div id="root"></div>
         <script type="module" src="/src/main.jsx"></script>
     </body>
