+++ src/components/Meals/MealItem/MealItem.jsx
@@ -1,35 +1,35 @@
-import { useContext } from 'react';
+import { useContext, useEffect, useState } from "react";
 
-import MealItemForm from './MealItemForm';
-import classes from './MealItem.module.css';
-import CartContext from '../../../store/CartContext';
+import MealItemForm from "./MealItemForm";
+import classes from "./MealItem.module.css";
+import CartContext from "../../../store/CartContext";
 
 const MealItem = (props) => {
-  const cartCtx = useContext(CartContext);
+    const cartCtx = useContext(CartContext);
 
-  const price = `$${props.price.toFixed(2)}`;
+    const price = `$${props.price.toFixed(2)}`;
 
-  const addToCartHandler = amount => {
-    cartCtx.addItem({
-      id: props.id,
-      name: props.name,
-      amount: amount,
-      price: props.price
-    });
-  };
+    const addToCartHandler = (amount) => {
+        cartCtx.addItem({
+            id: props.id,
+            name: props.name,
+            amount: amount,
+            price: props.price,
+        });
+    };
 
-  return (
-    <li className={classes.meal}>
-      <div>
-        <h3>{props.name}</h3>
-        <div className={classes.description}>{props.description}</div>
-        <div className={classes.price}>{price}</div>
-      </div>
-      <div>
-        <MealItemForm id={props.id} onAddToCart={addToCartHandler} />
-      </div>
-    </li>
-  );
+    return (
+        <li className={classes.meal}>
+            <div>
+                <h3>{props.name}</h3>
+                <div className={classes.description}>{props.description}</div>
+                <div className={classes.price}>{price}</div>
+            </div>
+            <div>
+                <MealItemForm id={props.id} onAddToCart={addToCartHandler} />
+            </div>
+        </li>
+    );
 };
 
 export default MealItem;
